Allow methods to be called prior to changing step.
Added onLeaving and onLeft methods to Step, which are called from the wizard during changeStep(). These allow methods - such a save - to be called on a step, and to check for errors before proceeding to the next step.

Changed getStep on wizard to return $this->steps, and to call $this->createSteps if it is empty.